+++ CHANGELOG.md
@@ -13,6 +13,7 @@ with the `YY.MINOR.MICRO` scheme.
 ### Features
 
 - Support unique indexes with non-sharded columns on sharded tables, with sharding key prefix provided
+- Introduce the governor script to add `is_default` column to the `_pico_tier` table
 - Pgproto now reports more verbose TLS error messages
 
 ### SQL